Add option to enable/disable hood entry
This adds a boolean option "active" to the hoods table that can be
used to hide a certain entry from geometry-based queries (e.g. when
adding it during preparation).

Queries based on the hood ID will still work even with active=0.

Note that this sets the default values for the DB to "0" as well.

Suggested-by: McUles <mcules@freifunk-hassberge.de>
Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>