Add option to enable/disable a polygon
This adds a boolean option "active" to the polyhoods table that can be
used to hide polygons from queries.

Note that since the polyhoods table relates hoods to polygons, setting
active=0 to a single entry there will disable that particular polygon,
while disabling the hood should be done in hoods table.

Note that a disabled polygon will be hidden in hoods.php overview,
while for the hoods themselves the active state is displayed.
Accordingly, map.html will show disabled hoods, but won't show
disabled polygons. This might be improved in a future patch ...

Note that this sets the default values for the DB to "0" as well.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>