Implement active hoods in voronoimap.php
This implements the active state for hoods in the voronoimap.php
by using different colors (blue=enabled, red=disabled) and only
enabling the active hoods by default.

The activation is implemented by checking the color for now, as
that's easier than creating an additional property for the active
state there.

Signed-off-by: Adrian Schmutzler <freifunk@adrianschmutzler.de>